Camera: Split camera model from camera
