+++ src/Game.cpp
@@ -21,12 +21,12 @@
 #include <spdlog/sinks/stdout_color_sinks.h>
 #include <spdlog/spdlog.h>
 
-#include "3D/Camera.h"
 #include "3D/CreatureBody.h"
 #include "3D/LandIslandInterface.h"
 #include "3D/Sky.h"
 #include "3D/Water.h"
 #include "Audio/AudioManagerInterface.h"
+#include "Camera/Camera.h"
 #include "Common/EventManager.h"
 #include "Common/RandomNumberManager.h"
 #include "Common/StringUtils.h"
